+++ lib/src/widget/pvc.dart
@@ -51,10 +51,9 @@ class MutablePVC<T extends ChangeNotifier> extends StatelessWidget {
   ///  * [StatefulWidget.initState] for more info.
   final _OnInit<T> onInit;
 
-  /// Creates a [StatelessWidget] using [ChangeNotifierProvider], and
-  /// bind the [controller] to it.
+  /// Create a widget using [ChangeNotifierProvider].
   ///
-  /// The parameters `controller` and `builder` must not be null.
+  /// The parameters [controller] and [builder] must not be null.
   const MutablePVC({
     Key key,
     @required this.controller,
@@ -123,9 +122,9 @@ class ImmutablePVC<T> extends StatelessWidget {
   ///  * [StatefulWidget.dispose] for more info.
   final Widget Function(BuildContext context, T controller) dispose;
 
-  /// Creates a [StatelessWidget] using [Provider], and bind the [controller] to it.
+  /// Create a widget using [Provider].
   ///
-  /// The parameters `controller` and `builder` must not be null.
+  /// The parameters [controller] and [builder] must not be null.
   const ImmutablePVC({
     Key key,
     @required this.controller,